ci: remove MsCoreUtils from Remotes
We now use the *devel* branch as development branch in all
RforMassSpectrometry projects. The `Remotes` field in the DESCRIPTION file
points to the *master* branch. Because we don't increment the version number
this forces a reinstallation of the *master* version after/on top of the *devel*
version (in fact a downgrade).